♻️ move coset and rowspace checks into mod2
